conversation_service_v2/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlparse==0.4.4

# Cache
redis==5.0.1

# AI/LLM
openai==1.12.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Monitoring
prometheus-client==0.19.0
